LayoutTests/inspector/worker/dom-debugger-url-breakpoints-expected.txt

Tests for URL breakpoint functionality (DOMDebugger) in a Worker.


== Running test suite: Worker.DOMDebugger.URLBreakpoint
-- Running test case: Worker.DOMDebugger.URLBreakpoint.Fetch
PASS: Should have paused.
PASS: Pause reason should have been Fetch.
PASS: Pause data URL should include "Fetch".
CALL STACK:
0: [F] loadResourceFetch
1: [F] triggerFetch
2: [F] handleMessage
ASYNC CALL STACK:
3: --- addEventListener ---
4: [P] Global Code

-- Running test case: Worker.DOMDebugger.URLBreakpoint.XHR
PASS: Should have paused.
PASS: Pause reason should have been XHR.
PASS: Pause data URL should include "XHR".
CALL STACK:
0: [F] loadResourceXHR
1: [F] triggerXHR
2: [F] handleMessage
ASYNC CALL STACK:
3: --- addEventListener ---
4: [P] Global Code
